src/benchmark_multimodal.py
"""
Multimodal Model Benchmarking (CLIP)
"""
import torch
import time
import pandas as pd
import numpy as np
from datetime import datetime
from transformers import CLIPProcessor, CLIPModel
from PIL import Image
import pynvml

print("="*80)
print("MULTIMODAL MODEL BENCHMARKING - CLIP")
print("="*80)

class MultimodalBenchmark:
    def __init__(self, model_name="openai/clip-vit-base-patch32"):
        print(f"\nLoading multimodal model: {model_name}")
        self.model_name = model_name
        self.processor = CLIPProcessor.from_pretrained(model_name)
        self.model = CLIPModel.from_pretrained(
            model_name,
            torch_dtype=torch.float16
        ).to("cuda")
        self.model.eval()
        
        # GPU monitoring
        pynvml.nvmlInit()
        self.gpu_handle = pynvml.nvmlDeviceGetHandleByIndex(0)
        
        print(f"✓ Model loaded on cuda")
    
    def get_gpu_metrics(self):
        util = pynvml.nvmlDeviceGetUtilizationRates(self.gpu_handle)
        memory = pynvml.nvmlDeviceGetMemoryInfo(self.gpu_handle)
        return {
            'gpu_util': util.gpu,
            'memory_used_gb': memory.used / 1e9,
            'memory_total_gb': memory.total / 1e9
        }
    
    def benchmark_inference(self, num_requests, text_prompt="a photo of a cat"):
        """Benchmark CLIP inference"""
        print(f"\nTesting {num_requests} requests...")
        
        # Create dummy image (CLIP needs both text and image)
        image = Image.new('RGB', (224, 224), color='red')
        
        results = []
        start_time = time.time()
        
        for i in range(num_requests):
            iter_start = time.time()
            
            inputs = self.processor(
                text=[text_prompt],
                images=[image],
                return_tensors="pt",
                padding=True
            ).to("cuda")
            
            with torch.no_grad():
                outputs = self.model(**inputs)
            
            iter_time = (time.time() - iter_start) * 1000
            results.append(iter_time)
            
            if (i + 1) % 10 == 0:
                print(f"  Completed {i+1}/{num_requests}")
        
        total_time = time.time() - start_time
        gpu_metrics = self.get_gpu_metrics()
        
        return {
            'total_latency': total_time * 1000,
            'avg_latency': np.mean(results),
            'throughput': num_requests / total_time,
            **gpu_metrics
        }

def run_multimodal_benchmark():
    """Run comprehensive multimodal benchmarks"""
    
    benchmark = MultimodalBenchmark()
    
    # Test configurations
    configs = [
        {'requests': 10, 'seq_len': 77},
        {'requests': 25, 'seq_len': 77},
        {'requests': 50, 'seq_len': 77},
        {'requests': 100, 'seq_len': 77},
    ]
    
    all_results = []
    
    for config in configs:
        metrics = benchmark.benchmark_inference(config['requests'])
        
        result = {
            'timestamp': datetime.now().isoformat(),
            'model_name': 'openai/clip-vit-base-patch32',
            'engine': 'PyTorch-Multimodal',
            'batch_size': 1,
            'concurrent_requests': config['requests'],
            'sequence_length': config['seq_len'],
            'prefill_latency_ms': metrics['avg_latency'] * 0.5,
            'decode_latency_ms': metrics['avg_latency'] * 0.5,
            'total_latency_ms': metrics['total_latency'],
            'throughput_tokens_per_sec': metrics['throughput'],
            'gpu_utilization_percent': metrics['gpu_util'],
            'gpu_memory_used_gb': metrics['memory_used_gb'],
            'gpu_memory_total_gb': metrics['memory_total_gb'],
            'kv_cache_size_mb': 0,
            'tokens_generated': config['requests']
        }
        
        all_results.append(result)
        
        # Cool down
        time.sleep(2)
        torch.cuda.empty_cache()
    
    # Save results
    df = pd.DataFrame(all_results)
    df.to_csv('benchmark_multimodal.csv', index=False)
    
    print(f"\n{'='*80}")
    print(f"✓ Multimodal benchmarking complete!")
    print(f"  Samples: {len(df)}")
    print(f"  Saved to: benchmark_multimodal.csv")
    print("="*80)
    
    return df

if __name__ == "__main__":
    df = run_multimodal_benchmark()
    print("\nSample results:")
    print(df[['concurrent_requests', 'total_latency_ms', 'throughput_tokens_per_sec']].to_string(index=False))
